Remove before_commit and rspec dev dependencies
Previously we had a model of using a gem we created called [Before_commit](https://github.com/EnvironmentAgency/before_commit) in our projects. It would run a series of checks at each commit to ensure the code being commited met various standards and agreed styles.

However we are moving to a model of instead having a central repo of linters and rule files, which would be used as part of the CI build for a project but locally it will be left to the developer to choose how they wish to check their submissions.

This commits includes the following changes

- Removes the **Before_commit** and **Rubocop** gems from the gemspec
- Corrects some of the gemspec info now that the project falls under the EA organisation
- Removes the calls to **Before_commit** and **Rubocop** in the Rakefile